% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/dimensions.R
\name{ExtendedDimensions}
\alias{ExtendedDimensions}
\title{GSDF extended dimensions}
\usage{
ExtendedDimensions(g)
}
\arguments{
\item{g}{Field to extract dimensions from}
}
\value{
Integer vector with indices of extended dimensions
}
\description{
Get the extended dimensions (those with length>1) from a field
}

