% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/rotate.R
\name{FieldToPole}
\alias{FieldToPole}
\title{Rotate a field}
\usage{
FieldToPole(g, pole.lat, pole.lon, greedy = FALSE)
}
\arguments{
\item{g}{field to be rotated}

\item{pole.lat}{latitude of pole to rotate to (degrees).}

\item{pole.lon}{longitude of pole to rotate to (degrees).}

\item{greedy}{boolean. How to deal with missing data: If FALSE (default)
missing regions will stay missing (and grow). If TRUE, missing
regions will shrink.}
}
\value{
input field but with the rotated pole.
}
\description{
Keeps the same grid but moves the North pole to a different place.
}
\details{
Effectively allows you to move the centre of the lat-lon grid (where
 distortion is low) to any position.
}
\seealso{
\code{\link{GSDF::RotateLatLon}} and \code{\link{GSDF::UnrotateLatLon}}.
}

