% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/dimensions.R
\name{FindDimension}
\alias{FindDimension}
\title{Find the index of a named dimension}
\usage{
FindDimension(g, label)
}
\arguments{
\item{g}{GSDF structure}

\item{label}{text string, 'lat','lon', 'ens' or 'time'}
}
\value{
Integer number of requested dimension, or NULL if not present.
}
\description{
Identify the dimension that contains 'lat', 'time' etc.
}
\details{
If the first dimension in g is 'lat', the second 'time', and
the third 'lon', \code{FindDimensions(g,'lon')} will
return 3.
}

