% Generated by roxygen2 (4.0.1): do not edit by hand
\name{GSDF.concatenate}
\alias{GSDF.concatenate}
\title{Merge two fields along one dimension.}
\usage{
GSDF.concatenate(d1, d2, dimn, fn)
}
\arguments{
\item{d1}{GSDF field}

\item{d2}{GSDF field}

\item{dimn}{name (e.g. 'ensemble') or number (3) of dimension to concatenate along}
}
\value{
reduced field.
}
\description{
Takes two n-dimensional fields where n-1 of the dimensions
 are identical and concatnate the fields along the remaining
 dimension.
}
\details{
If you have a 3-d field (say) lat*lon*time1, and asecond field
 lat*lon*time2 on the same lat lon grid - combine them to
 make 1 output field lat*lon*(time1+time2). Can concatenate
 along any dimension if all the other dimensions are the same.
}

