\name{GSDF.get.extended.dimensions}
\alias{GSDF.get.extended.dimensions}
\title{GSDF extended dimensions}
\usage{
  GSDF.get.extended.dimensions(g)
}
\arguments{
  \item{g}{Field to extract dimensions from}
}
\value{
  Integer vector with indices of extended dimensions
}
\description{
  Get the extended dimensions (those with length>1) from a
  field
}

