% Generated by roxygen2 (4.0.1): do not edit by hand
\name{GSDF.interp.surface}
\alias{GSDF.interp.surface}
\title{2d bilinear interpolation}
\usage{
GSDF.interp.surface(obj, loc.greedy = FALSE)
}
\arguments{
\item{obj}{A list with components x,y, and z in the same style as used
by contour, persp, image etc. x and y are the X and Y grid
values and z is a matrix with the corresponding values of the
surface.}

\item{loc}{A matrix of (irregular) locations to interpolate. First
column of loc is the X coordinates and second is the Y's.}

\item{greedy}{Boolean - if TRUE (default) NA are returned for regions of
the obj\$z where any neighbour points are NA (same as version in fields).
If FALSE, only return NA where all neighbouring points are NA.}
}
\value{
An vector of interpolated values.
}
\description{
Identical to interp.surface from the fields package. Except
 that it optionally ignores NA values in the source field.
}
\details{
Useful for regridding sea-ice and SST fields, which otherwise
 end up as missing along coastlines.
}

