% Generated by roxygen2 (4.0.2): do not edit by hand
\name{GSDF.ll.to.rg}
\alias{GSDF.ll.to.rg}
\title{Rotate lats and lons}
\usage{
GSDF.ll.to.rg(lat, lon, pole.lat, pole.lon, polygon = FALSE)
}
\arguments{
\item{lat}{vector of latitudes in standard pole (degrees).}

\item{lon}{vector of longitudes in standard pole (degrees).}

\item{pole.lat}{latitude of pole to rotate to (degrees).}

\item{pole.lon}{longitude of pole to rotate to (degrees).}

\item{polygon}{are lats and longs sets of polygon vertices
(probably from maps).}
}
\value{
list with components 'lat' and 'lon' - vectors of
  rotated lat and lon (in degrees).
}
\description{
From standard pole to same positions in a rotated pole.
}
\details{
Convert latitudes and longitudes to the equivalents
 with a rotated pole. (Formulae from UMDP S1).
}

