% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GSDF.ncdf.R
\name{GSDF.ncdf.write}
\alias{GSDF.ncdf.write}
\title{Write to netCDF file}
\usage{
GSDF.ncdf.write(f, file.name, name = "variable", missval = NULL,
  prec = "float", shuffle = FALSE, compression = NA, chunksizes = NA)
}
\arguments{
\item{f}{A GSDF field}

\item{file.name}{Output file name.}

\item{missval}{see \code{\link{ncvar_def}}}

\item{prec}{see \code{\link{ncvar_def}}}

\item{shuffle}{see \code{\link{ncvar_def}}}

\item{compression}{see \code{\link{ncvar_def}}}

\item{chunksizes}{see \code{\link{ncvar_def}}}
}
\value{
1 if successful, NULL for failure - invisibly.
}
\description{
Write a single field to a netCDF file.
Overwites anything already in the file.
}
\details{
Always uses 'hours since 1800-01-01:00' in the 
 gregorian calendar for time fields.
}
