% Generated by roxygen2 (4.0.1): do not edit by hand
\name{GSDF.print}
\alias{GSDF.print}
\title{Print a GSDF structure}
\usage{
GSDF.print(g)
}
\arguments{
\item{g}{GSDF structure to print}
}
\description{
\code{GSDF.print} prints a condensed version of the structure
 provided as its first argument.
}

