% Generated by roxygen2 (4.0.1): do not edit by hand
\name{GSDF.reduce.1d}
\alias{GSDF.reduce.1d}
\title{Remove a dimension from a field}
\usage{
GSDF.reduce.1d(d, dimn, fn, ...)
}
\arguments{
\item{d}{GSDF field}

\item{dimn}{name (e.g. 'ensemble') or number (3) of dimension to reduce}

\item{fn}{function to do the reduction (often mean or sd).}

\item{...}{additional arguments to fn (e.g. na.rm=T)}
}
\value{
reduced field.
}
\description{
Convert a n-dimensional field to an (n-1) dimensional field
by applying a function over the selected dimension.
}
\details{
If you have a 3-d field (say) lat*lon*ensemble member, convert it
 to a two d (lat*lon) field with the ensemble mean by running this with
 function 'mean' and dimension 'ensemble'. But you can use any function
 (that has a vector input and scalar output) over any dimension.
}

