% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GSDF.R
\name{GSDF.select.from.1d}
\alias{GSDF.select.from.1d}
\title{Select one value from one dimension}
\usage{
GSDF.select.from.1d(d, dimn, idx, ...)
}
\arguments{
\item{d}{GSDF field}

\item{dimn}{name (e.g. 'ensemble') or number (3) of dimension to reduce}

\item{idx}{which member of that dimension to keep.}

\item{...}{additional arguments to fn (e.g. na.rm=T)}
}
\value{
reduced field.
}
\description{
Convert a n-dimensional field to an (n-1) dimensional field
 by selecting a single value from one dimension.
}
\details{
This is a special case of \code{\link{GSDF.reduce.1d}}, but
a common one, and it does not use an
 'apply' loop - so it's faster.
}
