% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GSDF.time.R
\name{GSDF.time.check.calendar}
\alias{GSDF.time.check.calendar}
\title{Check that the calendar is supported}
\usage{
GSDF.time.check.calendar(calendar)
}
\arguments{
\item{date}{A celendar string}
}
\description{
Can do gregorian, 360_day and 365_day calendars.
}
\details{
365_day may be called 'noleap'
gregorian may be called 'standard' or
'proleptic_gregorian'.
}
