% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GSDF.time.R
\name{GSDF.time.check.time}
\alias{GSDF.time.check.time}
\title{Check that dates are meaningful}
\usage{
GSDF.time.check.time(date)
}
\arguments{
\item{date}{A GSDF time object}
}
\description{
Check that each string date is valid
 given the calendar.
}
