% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GSDF.time.R
\name{GSDF.time.difference}
\alias{GSDF.time.difference}
\title{Calculate difference (in minutes) between two dates}
\usage{
GSDF.time.difference(first, second)
}
\arguments{
\item{first}{GSDF date}

\item{second}{GSDF date (same no. of dates as first)}
}
\description{
Need this for interpolation in time
}
