% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GSDF.time.R
\name{GSDF.time.from.base.and.offset}
\alias{GSDF.time.from.base.and.offset}
\title{Convert base + offset into GSDF.time format}
\usage{
GSDF.time.from.base.and.offset(offset, base, units, calendar)
}
\arguments{
\item{offset}{numeric vector of time offsets.}

\item{base}{String time in format 'YYYY-MM-DD:HH:SS'
smaller components (s, min) may be missing.}

\item{units}{Units of the offset - string 'hours', 'days', ...}

\item{calendar}{One of 'gregorian', '360_day' and '365_day'}
}
\description{
NetCDF files store date:time as a base time and numeric
 offsets from that time given a calendar. The offsets
can be in units of days, hours, or minutes.
}
\details{
Convert these into GSDF.time format
 (Fractions of a minute are discarded).
}
