% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GSDF.time.R
\name{GSDF.time.increment}
\alias{GSDF.time.increment}
\title{Add a period to a GSDF.time}
\usage{
GSDF.time.increment(date, offset, units)
}
\arguments{
\item{date}{- GSDF.time}

\item{offset}{numeric vector of time offsets.}

\item{units}{Units of the offset - string 'hours', 'days', ...}
}
\description{
Increment the time by a number of years, month, days,
 hours or minutes.
}
\details{
Increments can be negative, and can be fractional.
}
