% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GSDF.time.R
\name{GSDF.time}
\alias{GSDF.time}
\title{Unfortunately, the data we are interested in use several calendars.}
\usage{
GSDF.time(string, calendar)
}
\arguments{
\item{string}{Vector of date strings in format 'YYYY-MM-DD:HH:SS'}

\item{calendar}{One of 'gregorian', '360_day' and '365_day'}
}
\description{
The 'standard' or 'gregorian' calendar
The 'proleptic gregorian' calendar
A '360_day' calendar with 12 months each of 30 days
A '365_day' or 'noleap' calendar with 12 months of
  c(31,28,31,30,31,30,31,31,30,31,30,31) days.
}
\details{
So we need custom date tools - we will store a date as an
 ISO-8601 format string down to minutes: 'YYYY-MM-DD:HH:MM'
 combined with a string describing the calendar in use.
Fortunately, we don't need to bother with time zones.
Negative years are not currently supported.
}
