% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GSDF.time.R
\name{GSDF.time.recalendar}
\alias{GSDF.time.recalendar}
\title{Adjust time values for change in calendar}
\usage{
GSDF.time.recalendar(date, new.calendar)
}
\arguments{
\item{first}{GSDF date}

\item{second}{GSDF date (same no. of dates as first)}
}
\description{
Sometimes we want to specify a time range, but we don't know the calendar
e.g. a year is -01-01:00:00 to -12-31:23:59 in Gregorian and
               -01-01:00:00 to -12:30:23:59 in 360_day.
So it's useful to give the spec in gregorian and auto-shrink it if the
 data turns out to have a 360_day calendar when we open the file.
}
\details{
Assumes that you are going for the full month or year - use
 with caution.
}
