% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GSDF.R
\name{GSDF.unleap}
\alias{GSDF.unleap}
\title{Take out data for February 29th.}
\usage{
GSDF.unleap(d)
}
\arguments{
\item{d}{GSDF field}
}
\value{
field with Feb 29th removed
}
\description{
Removes all data for leap-days from a GSDF field.
 Has no effect if input field has no data from Feb 29th.
}
\details{
Often want to compare data from two years, one of which was a leap year
 simplest approach is to strip Feb 29th data before comparison.
Most often occurs when making anomalies.
}
