% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/interpolate.R
\name{InterpolateLatLon}
\alias{InterpolateLatLon}
\title{Interpolate lat:lon points in a 2d field}
\usage{
InterpolateLatLon(g, lat, lon, full = NULL, greedy = FALSE)
}
\arguments{
\item{g}{GSDF data structure (must have lat and lon as dimensions)}

\item{lat}{latitudes to interpolate to}

\item{lon}{longitude values to interpolate to}

\item{full}{- TRUE => apply appropriate boundary conditions for full
global lat:lon field (wrap longitudes and extrapolate near poles if required,
FALSE => just interpolate (removes data outside range of g),
NULL (default) => Guess which to do (is field ll and covering most of globe?).}

\item{greedy}{boolean. If FALSE (default) NA values in the input field
propagate to generate NA values in all output values adjacent to them.
If true, output values are NA only if all adjacent grid points are NA.}
}
\value{
numeric vector of interpolated values
}
\description{
2d bilinear interpolation
}
\details{
Basically the same as \code{\link{Interpolate2d}} except
that we check that the field has lat and lon as dimensions
and arrange them in the right order.
}

