% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/utils.R
\name{print.GSDF}
\alias{print.GSDF}
\title{Print a GSDF structure}
\usage{
\method{print}{GSDF}(g)
}
\arguments{
\item{g}{GSDF structure to print}
}
\description{
\code{GSDF.print} prints a condensed version of the structure
 provided as its first argument.
}

