% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/dimensions.R
\name{RollDimensions}
\alias{RollDimensions}
\title{Roll-out a pair of dimensions}
\usage{
RollDimensions(g, d1, d2)
}
\arguments{
\item{g}{GSDF data structure}

\item{d1}{name or integer index of first dimension}

\item{d2}{name or integer index of second dimension}
}
\value{
numeric vector of first dimension values
}
\description{
Get (for example) lat for each point in a 2d slice
}
\details{
If g has 2 dimensions, 'lat' and 'lon', each of length 10,
\code{RollDimensions(g,1,2)} will return the latitude
for each of the 100 points in the slice; and
\code{RollDimensions(g,2,1)} will return the longitude
similarly.
}

