% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/leap_year.R
\name{Unleap}
\alias{Unleap}
\title{Take out data for February 29th.}
\usage{
Unleap(d)
}
\arguments{
\item{d}{GSDF field}
}
\value{
field with Feb 29th removed
}
\description{
Removes all data for leap-days from a GSDF field.
 Has no effect if input field has no data from Feb 29th.
}
\details{
Often want to compare data from two years, one of which was a leap year
 simplest approach is to strip Feb 29th data before comparison.
Most often occurs when making anomalies.
}

